Application form be an independent component imported in job detail
