+++ frontend/src/components/ApplicationForm.css
@@ -1,0 +1,58 @@
+  .application-form {
+    background-color: #fff;
+    padding: 20px;
+    border-radius: 8px;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+    position: sticky;
+    top: 20px;
+  }
+  
+  .application-form form {
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
+  }
+  
+  .application-form label {
+    font-weight: 500;
+    color: #333;
+  }
+  
+  .application-form input,
+  .application-form select,
+  .application-form textarea {
+    padding: 10px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+    font-size: 1em;
+  }
+  
+  .application-form textarea {
+    resize: vertical;
+    min-height: 100px;
+  }
+  
+  .application-form button {
+    background-color: #2196f3;
+    color: white;
+    padding: 12px;
+    border: none;
+    border-radius: 4px;
+    font-size: 1.1em;
+    cursor: pointer;
+    transition: background-color 0.3s;
+  }
+  
+  .application-form button:hover {
+    background-color: #1976d2;
+  }
+  
+  @media (max-width: 768px) {
+    .job-info-container {
+      grid-template-columns: 1fr;
+    }
+    
+    .application-form {
+      position: static;
+    }
+  }
